Version 2.1
- Improved file structure to separate function access between user and developer code (grib_sections.cpp, grib_sections.hpp)
- Added option to run the program with command-line arguments
- Minor changes in function structure
