+++ OUT.txt
@@ -1,0 +1,30 @@
+File Opened: 20150310.00.W.dwa_griby.grib
+-------------Section 0-------------
+Length of file: 4538
+GRIB version: 1
+
+-------------Section 1-------------
+Length of Section: 28
+GDS: Included
+BMS: Omitted
+Time of forecast: 2015-03-10 00:00
+Scale factor D: 1
+
+-------------Section 2-------------
+Length of Section: 178
+Latitude start: 0
+Longitude start: -30
+Latitude end: 90
+Longitude end: 60
+Bytes 29-33 (decimal): 0 0 0 0 0 
+
+-------------Section 4-------------
+Length of Section: 4320
+Scale factor E: 0
+Reference value: -102
+Number of bits into which a datum point is packed: 10
+Bytes 12-16 (decimal): 31 72 82 8 136 
+
+-------------GRIB Length Check-------------
+Declared GRIB message length: 4538
+Sum of section lengths:       4526
